lab_6a/pyproject.toml
[project]
name = "lab_6a"
version = "0.1.0"
description = "QLoRA Vision Finetuning (Qwen3-VL-8B LaTeX OCR)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "comet_ml",
    "unsloth",
    "transformers",
    "datasets",
    "trl",
    "huggingface_hub",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = []
